+++ lib/openapi_first/test/coverage/route_task.rb
@@ -4,8 +4,30 @@ module OpenapiFirst
   module Test
     module Coverage
       RouteTask = Data.define(:path, :request_method, :requests, :responses) do
+        def skipped?
+          requests.all?(&:skipped?)
+        end
+
+        def tracked_requests
+          requests.reject(&:skipped?)
+        end
+
+        def tracked_responses
+          responses.reject(&:skipped?)
+        end
+
+        def skipped_requests
+          requests.select(&:skipped?)
+        end
+
+        def skipped_responses
+          responses.select(&:skipped?)
+        end
+
         def finished?
-          requests.all?(&:finished?) && responses.all?(&:finished?)
+          return false if skipped?
+
+          tracked_requests.all?(&:finished?) && tracked_responses.all?(&:finished?)
         end
 
         def summary